moved crypto section to ee
